ComputerVisionInterface/requirements.txt
opencv_python_headless==4.5.2.54
streamlit==0.82.0
mediapipe==0.8.4.2
numpy==1.22.0
Pillow>=9.2.0
